+++ docs/changelog.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add `pyanalyze.extensions.NoReturnGuard` (#370)
+- Support configuration in a `pyproject.toml` file (#368)
 - Require `typeshed_client` 2.0 (#361)
 - Add JSON output for integrating pyanalyze's output with other
   tools (#360)